Refactor into PaperVersion
