+++ components/Document/lib/PaperVersion/PaperVersionAuthorsAndMetadataStep.tsx
@@ -1,0 +1,12 @@
+import Button from "~/components/Form/Button";
+import { Paper } from "../types";
+
+const PaperVersionAuthorsAndMetadataStep = ({ paper }: { paper: Paper | null }) => {
+  return (
+    <div>
+      TEST
+    </div>
+  );
+}
+
+export default PaperVersionAuthorsAndMetadataStep;
