+++ components/Document/lib/PaperVersion/PaperVersionContentStep.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import BaseModal from "~/components/Modals/BaseModal";
 import { StyleSheet, css } from "aphrodite";
-import { fetchDocumentByType } from "./fetchDocumentByType";
-import { DocumentVersion, Paper, parsePaper, WORK_TYPE } from "./types";
+import { fetchDocumentByType } from "../fetchDocumentByType";
+import { DocumentVersion, Paper, parsePaper, WORK_TYPE } from "../types";
 import FormInput from "~/components/Form/FormInput";
 import FormSelect from "~/components/Form/FormSelect";
 import FormTextArea from "~/components/Form/FormTextArea";
@@ -11,13 +11,6 @@ import HubSelectDropdown, {
 } from "~/components/Hubs/HubSelectDropdown";
 import { Hub } from "~/config/types/hub";
 import Button from "~/components/Form/Button";
-
-
-interface Args {
-  isOpen: boolean;
-  closeModal: () => void;
-  versions: DocumentVersion[];
-}
 
 export const articleTypeOptions: Array<{ value: WORK_TYPE; label: string }> = [
   {
@@ -30,75 +23,7 @@ export const articleTypeOptions: Array<{ value: WORK_TYPE; label: string }> = [
   },
 ];
 
-export type STEP =
-  | "CONTENT"
-  | "AUTHORS_AND_METADATA"
-  | "DECLARATIONS"
-  | "PREVIEW";
-
-export const ORDERED_STEPS: STEP[] = [
-  "CONTENT",
-  "AUTHORS_AND_METADATA",
-  "DECLARATIONS",
-  "PREVIEW",
-];
-
-const PaperVersionModal = ({ isOpen, closeModal, versions }: Args) => {
-  const [step, setStep] = useState<STEP>("CONTENT");
-  const [latestPaper, setLatestPaper] = useState<Paper | null>(null);
-  const handleNext = () => {
-    const currentIndex = ORDERED_STEPS.indexOf(step);
-    if (currentIndex + 1 < ORDERED_STEPS.length) {
-      setStep(ORDERED_STEPS[currentIndex + 1]);
-    }
-  };
-
-  useEffect(() => {
-    (async () => {
-      const latestVersion = versions.filter((version) => version.isLatest)[0];
-      const rawPaper = await fetchDocumentByType({
-        documentType: "paper",
-        documentId: latestVersion.id,
-      });
-      const parsed = parsePaper(rawPaper);
-      setLatestPaper(parsed);
-    })();
-  }, []);
-
-  return (
-    <BaseModal
-      isOpen={isOpen}
-      hideClose={false}
-      closeModal={() => {
-        closeModal();
-      }}
-      zIndex={1000000}
-      title={"Submit new version"}
-      modalContentStyle={styles.modalStyle}
-    >
-      <div className={css(styles.modalBody)}>
-        {step === "CONTENT" && <ContentStep paper={latestPaper} />}
-        {step === "AUTHORS_AND_METADATA" && <AuthorsAndMetadataStep paper={latestPaper} />}
-      </div>
-
-
-      <div className={css(formStyles.buttonWrapper)}>
-        <Button
-          label={"Next"}
-          onClick={() => handleNext()}
-          theme="solidPrimary"
-        />
-      </div>
-
-    </BaseModal>
-  );
-};
-
-const AuthorsAndMetadataStep = ({ paper }: { paper: Paper | null }) => {
-  return <div>Authors and metadata step</div>;
-}
-
-const ContentStep = ({ paper }: { paper: Paper | null }) => {
+const PaperVersionContentStep = ({ paper }: { paper: Paper | null }) => {
   const [title, setTitle] = useState<null | string>(null);
   const [selectedWorkType, setSelectedWorkType] =
     useState<WORK_TYPE>("article");
@@ -193,14 +118,4 @@ const formStyles = StyleSheet.create({
   }
 });
 
-const styles = StyleSheet.create({
-  modalBody: {
-    width: "100%",
-  },
-  modalStyle: {
-    width: 650,
-    minHeight: 500,
-  },
-});
-
-export default PaperVersionModal;
+export default PaperVersionContentStep;
