+++ components/Document/lib/PaperVersion/PaperVersionTypes.ts
@@ -1,0 +1,12 @@
+export type STEP =
+  | "CONTENT"
+  | "AUTHORS_AND_METADATA"
+  | "DECLARATIONS"
+  | "PREVIEW";
+
+export const ORDERED_STEPS: STEP[] = [
+  "CONTENT",
+  "AUTHORS_AND_METADATA",
+  "DECLARATIONS",
+  "PREVIEW",
+];
